security-manager-main/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
celery==5.3.6
redis==5.0.1
tortoise-orm==0.20.0
asyncpg==0.29.0
aerich==0.7.2  # For migrations
python-multipart==0.0.6
requests==2.31.0
PyYAML==6.0.1
semgrep==1.59.0  # For analysis node
google-generativeai==0.8.3
toons==0.5.2
